requirements.txt
# Core Azure packages
azure-functions
azure-cosmos
azure-ai-formrecognizer
azure-core
azure-identity

# OpenAI and environment
openai
python-dotenv

# HTTP and async
aiohttp
requests

# Data processing
pandas
numpy
python-dateutil

# Additional dependencies
certifi
typing-extensions
charset-normalizer
idna
multidict
yarl
six
urllib3
attrs